temp moved -- there are still thing left to check and evaluate
